+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Changelog
 asyncactor
 ++++++++++
 
+0.20.0
+------
+
+- introduced versioned parameters
+- introduced typed messages
+
 0.1.0
 -----